admin panel methods

+++ routes/web.php
@@ -13,20 +13,19 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::group(['namespace' => 'Post'], function (){
+Route::group(['namespace' => 'Post'], function () {
     Route::get('/posts', 'IndexController')->name('post.index');
-    Route::get('/posts/create', 'CreateController')->name('post.create');
-
-    Route::post('/posts', 'StoreController')->name('post.store');
     Route::get('/posts/{post}', 'ShowController')->name('post.show');
-    Route::get('/posts/{post}/edit', 'EditController')->name('post.edit');
-    Route::patch('/posts/{post}', 'UpdateController')->name('post.update');
-    Route::delete('/posts/{post}', 'DestroyController')->name('post.delete');
 });
 
 
 Route::group(['namespace' => 'Admin', 'prefix' => 'admin'], function () {
     Route::group(['namespace' => 'Post'], function () {
         Route::get('/post', 'IndexController')->name('admin.post.index');
+        Route::delete('/posts/{post}', 'DestroyController')->name('admin.post.delete');
+        Route::get('/posts/{post}/edit', 'EditController')->name('admin.post.edit');
+      Route::patch('/posts/{post}', 'UpdateController')->name('admin.post.update');
+//      Route::get('/posts/create', 'CreateController')->name('post.create');
+//      Route::post('/posts', 'StoreController')->name('post.store');
     });
 });
